Revert "transacao saida"
