+++ ProjetoLojaDesktop/ProjetoLojaDesktop/App.Config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <connectionStrings>
-    <add name="ProjetoLojaEntities" connectionString="metadata=res://*/Entity.ProjetoLojaEntity.csdl|res://*/Entity.ProjetoLojaEntity.ssdl|res://*/Entity.ProjetoLojaEntity.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=AZUL44\MSSQLSERVER2;initial catalog=ProjetoLoja;integrated security=True;multipleactiveresultsets=True;App=EntityFramework&quot;" providerName="System.Data.EntityClient" />
+    <add name="ProjetoLojaEntities" connectionString="metadata=res://*/Entity.ProjetoLojaEntity.csdl|res://*/Entity.ProjetoLojaEntity.ssdl|res://*/Entity.ProjetoLojaEntity.msl;provider=System.Data.SqlClient;provider connection string=&quot;Data Source=.\SQLEXPRESS;Initial Catalog=ProjetoLoja;Integrated Security=True;MultipleActiveResultSets=True&quot;" providerName="System.Data.EntityClient" />
   </connectionStrings>
 </configuration>
