--- ProjetoLojaDesktop/ProjetoLojaDesktop/FormFornecedor.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-using ProjetoLojaDesktop.Forms;
-
-namespace ProjetoLojaDesktop
-{
-    public partial class FormFornecedor : Form
-    {
-        public FormFornecedor()
-        {
-            InitializeComponent();            
-        }
-    }
-}
