--- ProjetoLojaDesktop/ProjetoLojaDesktop/FormPrincipal.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace ProjetoLojaDesktop
-{
-    public partial class FormPrincipal : Form
-    {
-        public FormPrincipal()
-        {
-            InitializeComponent();
-        }
-    }
-}
