--- ProjetoLojaDesktop/ProjetoLojaDesktop/Forms/FormMarca.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace ProjetoLojaDesktop.Forms
-{
-    public partial class FormMarca : Form
-    {
-        public FormMarca()
-        {
-            InitializeComponent();
-        }
-
-        private void btnLimpar_Click(object sender, EventArgs e)
-        {
-
-        }
-
-
-    }
-}
