--- ProjetoLojaDesktop/ProjetoLojaDesktop/Forms/FormModelo.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace ProjetoLojaDesktop.Forms
-{
-    public partial class FormModelo : Form
-    {
-        public FormModelo()
-        {
-            InitializeComponent();
-        }
-    }
-}
