--- ProjetoLojaDesktop/ProjetoLojaDesktop/Forms/Frm_cadCliente.cs
@@ -1,130 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Windows.Forms;
-
-namespace ProjetoLojaVirtual
-{
-    public partial class Frm_cadCliente : Form
-    {
-        public Frm_cadCliente()
-        {
-            InitializeComponent();
-        }
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void button3_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label1_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label2_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void textBox3_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label3_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void textBox2_TextChanged(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label6_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void panel1_Paint(object sender, PaintEventArgs e)
-        {
-
-        }
-
-        private void label7_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label14_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void aba_endereco_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label16_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void button5_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btn_adicionar_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btn_alterar_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btn_excluir_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btn_cancelar_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void btn_salvar_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void cbx_tipoPessoa_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            if (cbx_tipoPessoa.SelectedIndex == 0)
-            {
-                txt_cpf.Mask = "999,999,999-99";
-            }
-            else {
-                txt_cpf.Mask = "99,999,999/9999-99";
-            }
-        }
-
-        private void txt_telefone_MaskInputRejected(object sender, MaskInputRejectedEventArgs e)
-        {
-
-        }
-    }
-}
