+++ ProjetoLojaDesktop/ProjetoLojaDesktop/Program.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 using ProjetoLojaDesktop.Forms;
+using ProjetoLojaDesktop.Views;
+using ProjetoLojaDesktop;
 
 
 namespace cad_fornecedor
@@ -17,7 +19,7 @@ namespace cad_fornecedor
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new FormTransacaoSaida());
+            Application.Run(new FormPrincipal());
         }
     }
 }
